requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
feedparser>=6.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tzdata>=2024.1
